$value is not a instance of checkbox
I were having issue with BlockType.
when I request for query 
 ...on document_editableBlock{
          _editableName
          _editableType
          indices
        }
it return null value of editable name and type. but when removed condition of instance of checkbox. it started working. So I am proposing this change to plugin.
Because I don't want to change in my vendor folder. it will create problem while up-gradation. So If this change find correct then please merge.